Refactor ArtFile code
 - Create a using for Palette
 - move Color struct to its own file
 - Create a Sprite ImageType enum
