+++ OP2Utility.vcxproj.filters
@@ -147,6 +147,12 @@
     <ClInclude Include="src\Sprites\SectionHeader.h">
       <Filter>Sprites</Filter>
     </ClInclude>
+    <ClInclude Include="src\Sprites\Color.h">
+      <Filter>Sprites</Filter>
+    </ClInclude>
+    <ClInclude Include="src\Sprites\ImageType.h">
+      <Filter>Sprites</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="src\XFile.cpp">
