+++ src/Sprites/ArtFile.cpp
@@ -15,4 +15,3 @@ void ArtFile::ValidateImageMetadata() const
 		}
 	}
 }
-
